refactor(api): extract manual input formulas into new file backend/formulas.ts
